backend/requirements.txt
# =====================================================
# DIAGNOVET - DEPENDENCIAS DEL BACKEND
# =====================================================

# Framework web
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Base de datos y ORM
sqlalchemy==2.0.34
pydantic==2.11.9
supabase==2.19.0

# Configuración y variables de entorno
python-dotenv==1.0.0

# Procesamiento de PDFs e imágenes
Pillow==10.4.0
pytesseract==0.3.13
pdf2image==1.17.0
pdfjs-dist==3.11.174

# Análisis de datos
pandas==2.3.2
numpy==1.26.4

# APIs externas
openai==1.3.7
httpx==0.28.1

# Google APIs
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-auth==2.23.4
google-generativeai==0.8.3

# Autenticación y seguridad
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0

# Utilidades
uuid==1.30
